tests/wpt/web-platform-tests/css/css-grid-1/alignment/grid-self-alignment-stretch-vertical-rl-013.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Grid Layout Test: Self-Alignment and stretch on auto-sized tracks</title>
<link rel="author" title="Javier Fernandez Garcia-Boente" href="mailto:jfernandez@igalia.com">
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-align-self">
<link rel="help" href="https://drafts.csswg.org/css-align-3/#propdef-justify-self">
<link rel="help" href="https://drafts.csswg.org/css-align/#valdef-justify-self-stretch">
<link rel="help" href="https://drafts.csswg.org/css-align/#valdef-align-self-stretch">
<meta name="assert" content="Orthogonal grid items with 'stretch' value for align-self and/or justify-self are stretched along the column and/or row axis respectively; they will be sized as fit-content otherwise.">
<style>
.grid {
  position: relative;
  display: inline-grid;
  font: 20px/1 Ahem;
  background: grey;
  width: 250px;
  height: 250px;
  writing-mode: vertical-rl;
}
.grid > div { writing-mode: horizontal-tb; }
.firstRowFirstColumn {
  grid-row: 1;
  grid-column: 1;
  background: green;
  justify-self: stretch;
  align-self: start;
}
.firstRowSecondColumn {
  grid-row: 1;
  grid-column: 2;
  background: blue;
  justify-self: start;
  align-self: stretch;
}
.secondRowFirstColumn {
  grid-row: 2;
  grid-column: 1;
  background: yellow;
  justify-self: start;
  align-self: start;
}
.secondRowSecondColumn {
  grid-row: 2;
  grid-column: 2;
  background: red;
  justify-self: stretch;
  align-self: stretch;
}
</style>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../support/check-layout-th.js"></script>
<body onload="checkLayout('.grid')">
<div class="grid">
  <div data-offset-x="170" data-offset-y="0"   data-expected-width="80"  data-expected-height="125" class="firstRowFirstColumn">XX X</div>
  <div data-offset-x="125" data-offset-y="125" data-expected-width="125" data-expected-height="20"  class="firstRowSecondColumn">XX X</div>
  <div data-offset-x="45"  data-offset-y="0"   data-expected-width="80"  data-expected-height="20"  class="secondRowFirstColumn">XX X</div>
  <div data-offset-x="0"   data-offset-y="125" data-expected-width="125" data-expected-height="125" class="secondRowSecondColumn">XX X</div>
</div>
